+++ crates/app-data/tests/common/mod.rs
@@ -9,10 +9,6 @@ pub const APP_DATA_HEX_2: &str =
     "0x8af4e8c9973577b08ac21d17d331aade86c11ebcc5124744d621ca8365ec9424";
 pub const CID_2: &str = "f01551b208af4e8c9973577b08ac21d17d331aade86c11ebcc5124744d621ca8365ec9424";
 
-pub const APP_DATA_HEX_LEGACY: &str =
-    "0x447320af985c5e834321dc495545f764ad20d8397eeed2f4a2dcbee44a56b725";
-pub const CID_LEGACY: &str = "QmSwrFbdFcryazEr361YmSwtGcN4uo4U5DKpzA4KbGxw4Q";
-
 pub const PINATA_IPFS_HASH: &str = "QmU4j5Y6JM9DqQ6yxB6nMHq4GChWg1zPehs1U7nGPHABRu";
 pub const PINATA_APP_DATA_HEX: &str =
     "0x5511c4eac66ab272d9a6ab90e07977d00ff7375fc4dc1038a3c05b2c16ca0b74";
@@ -20,8 +16,6 @@ pub const PINATA_APP_DATA_HEX: &str =
 pub const APP_DATA_STRING: &str =
     "{\"appCode\":\"CoW Swap\",\"metadata\":{},\"version\":\"0.7.0\"}";
 pub const APP_DATA_STRING_2: &str = "{\"appCode\":\"CoW Swap\",\"environment\":\"production\",\"metadata\":{\"quote\":{\"slippageBips\":\"50\",\"version\":\"0.2.0\"},\"orderClass\":{\"orderClass\":\"market\",\"version\":\"0.1.0\"}},\"version\":\"0.6.0\"}";
-pub const APP_DATA_STRING_LEGACY: &str =
-    "{\"version\":\"0.7.0\",\"appCode\":\"CowSwap\",\"metadata\":{}}";
 
 pub fn parity_fixture() -> Value {
     serde_json::from_str(include_str!("../../../../parity/fixtures/app-data.json")).unwrap()